Enable fast page-stepping on block pages
